requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
dj-database-url==2.1.0
Django==4.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-ckeditor==6.7.0
django-js-asset==2.1.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
